requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
python-owasp-zap-v2.4>=0.0.20
wfuzz>=3.1.0
python-nmap>=0.7.1
click>=8.1.7
rich>=13.6.0
pyyaml>=6.0.1
markdown>=3.5
flask>=3.0.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.5
python-dotenv>=1.0.0
gunicorn>=21.2.0
jsonschema>=4.20.0
cvss>=2.6